Fix bug that allowed frame_size == 0.
In the rare case that a prescan exposure generates fewer SE than it destroys, the frame_size could be zero. This would lead to an infinite loop for the accumulator. Worse, in the real exposure no actual iteration would be executed.

In any case, a frame_size of zero does not make sense. So set it to 1 at least.